Develop

+++ test.py
@@ -1,26 +1,52 @@
 import os
+import io
+import codecs
+import difflib
 import xml.dom.minidom as dom
-import sqlite3 as db
 from modules.xml_to_ram import XMLtoRAM
-from modules.ram_to_dbd import RAMToDBD
+from modules.ram_to_dbd import RAMtoDBD
+from modules.dbd_to_ram import DBDtoRAM
+from modules.ram_to_xml import RAMtoXML
 
-xml = dom.parse(os.path.join("xml/", "xml_test.xml"))
-ram = XMLtoRAM(xml).xml_to_ram()
-RAMToDBD(ram, "test.db").ram_to_dbd()
+if os.path.exists("diff_error.html"):
+    os.remove("diff_error.html")
+for t in ["tasks", "prjadm"]:
+    if os.path.exists("test.xml"):
+        os.remove("test.xml")
+    if os.path.exists("test.db"):
+        os.remove("test.db")
+    if os.path.exists("diff_"+t+".html"):
+        os.remove("diff_"+t+".html")
 
-db = db.connect("test.db")
-cursor = db.cursor()
-go = '1'
-while go == '1':
-    print('type?')
-    c = input()
-    if c == 'schemas' or c == 'domains' or c == 'tables' or c == 'fields' or c == 'constraints' or c == 'indices':
-        c = 'select * from dbd$'+c
-        cursor.execute(c)
-        t = cursor.fetchall()
-        for tt in t:
-            print(tt)
-    print('next?')
-    go = input()
-db.close()
-os.remove("test.db")
+    xml_file = os.path.join("xml/", t+".xml")
+
+    xml = dom.parse(xml_file)
+    ram = XMLtoRAM(xml).xml_to_ram()
+    RAMtoDBD(ram, "test.db").ram_to_dbd()
+    ram2 = DBDtoRAM("test.db").schema()
+    xml2 = RAMtoXML(ram2).ram_to_xml()
+
+    with open("test.xml", "wb") as f:
+        f.write(xml2.toprettyxml(encoding="utf-8", indent="  "))
+
+    text1 = io.open(xml_file, encoding='utf-8').readlines()
+    text2 = io.open("test.xml", encoding='utf-8').readlines()
+    d = difflib.HtmlDiff()
+    html = d.make_file(text1, text2)
+    file = codecs.open("diff_"+t+".html", 'w', 'utf-8')
+    file.write(html)
+    file.close()
+# make fake error file
+for table in ram2.tables:
+    for field in table.fields:
+        field.name = "error"
+xml2 = RAMtoXML(ram2).ram_to_xml()
+with open("test.xml", "wb") as f:
+    f.write(xml2.toprettyxml(encoding="utf-8", indent="  "))
+text1 = io.open(xml_file, encoding='utf-8').readlines()
+text2 = io.open("test.xml", encoding='utf-8').readlines()
+d = difflib.HtmlDiff()
+html = d.make_file(text1, text2)
+file = codecs.open("diff_error.html", 'w', 'utf-8')
+file.write(html)
+file.close()
